+++ workspaces/packages/bicep-assets/src/commands/deploy.ts
@@ -1,5 +1,5 @@
-import { DefaultAzureCredential } from '@azure/identity';
-import axios from 'axios';
+import { AzureCliCredential } from '@azure/identity';
+import { BlobServiceClient } from '@azure/storage-blob';
 import { Command, Option } from 'clipanion';
 import { mkdtemp, readFile, rm } from 'fs/promises';
 import { existsSync, statSync } from 'fs';
@@ -8,12 +8,20 @@ import { join } from 'path';
 import * as t from 'typanion';
 import { archiveFolder } from 'zip-lib';
 
+import { Configuration } from '../configuration';
 import { isManifest } from '../utils/manifest';
 
 export class DeployCommand extends Command {
   static paths = [
     ['deploy'],
   ];
+
+  static usage = Command.Usage({
+    description: 'Deploys the assets to the storage account',
+    details: `
+      This command will deploy the assets in the dist folder to the storage account.
+    `,
+  });
 
   distFolder = Option.String('--dist-folder', '.bicep-assets');
 
@@ -34,47 +42,26 @@ export class DeployCommand extends Command {
 
     t.assert(manifestContent, isManifest);
 
-    const files = Object.values(manifestContent.assets).map(fileName => {
-      const stat = statSync(join(this.distFolder, fileName));
-      return stat.isDirectory() ? `${fileName}.zip` : fileName;
-    });
-
-    const uploads = await this.getUploads(files, manifestContent.resourceProviderId);
-
-    console.log(uploads);
 
     for (const [name, fileName] of Object.entries(manifestContent.assets)) {
       const stat = statSync(join(this.distFolder, fileName));
       const targetFileName = stat.isDirectory() ? `${fileName}.zip` : fileName;
 
-      if (uploads[targetFileName]) {
-        await this.uploadAsset(name, fileName, uploads[targetFileName], targetFileName, stat.isDirectory());
-      }
+      await this.uploadAsset(name, fileName, targetFileName, stat.isDirectory());
     }
   }
 
-  async getUploads(assetFiles: string[], resourceProviderId: string): Promise<Record<string, string>> {
-    const creds = new DefaultAzureCredential();
-    const token = await creds.getToken(['https://management.azure.com/.default']);
+  async uploadAsset(name: string, fileName: string, targetFileName: string, compress: boolean) {
+    const configuration = await Configuration.load(false);
+    const creds = new AzureCliCredential();
 
-    const res = await axios.post(`https://management.azure.com${resourceProviderId}/uploadAssetsSas`, {
-      properties: {
-        assets: assetFiles,
-      },
-    }, {
-      params: {
-        // eslint-disable-next-line @typescript-eslint/naming-convention
-        'api-version': '2018-09-01-preview',
-      },
-      headers: {
-        Authorization: `Bearer ${token.token}`,
-      },
-    });
+    const blobServiceClient = new BlobServiceClient(
+      `https://${configuration.storageAccountName}.blob.core.windows.net`,
+      creds,
+    );
 
-    return res.data.assets;
-  }
+    const containerClient = blobServiceClient.getContainerClient('assets');
 
-  async uploadAsset(name: string, fileName: string, sasUrl: string, targetFileName: string, compress: boolean) {
     const tempFolder = await mkdtemp(tmpdir());
     try {
       console.log(`Publishing asset: ${fileName} (${name})`);
@@ -82,20 +69,11 @@ export class DeployCommand extends Command {
         await archiveFolder(join(this.distFolder, fileName), join(tempFolder, targetFileName));
 
         const zipFile = await readFile(join(tempFolder, targetFileName));
-        await axios.put(sasUrl, zipFile, {
-          headers: {
-            // eslint-disable-next-line @typescript-eslint/naming-convention
-            'x-ms-blob-type': 'BlockBlob',
-          },
-        });
+
+        containerClient.uploadBlockBlob(targetFileName, zipFile, zipFile.length);
       } else {
         const content = await readFile(join(this.distFolder, fileName));
-        await axios.put(sasUrl, content, {
-          headers: {
-            // eslint-disable-next-line @typescript-eslint/naming-convention
-            'x-ms-blob-type': 'BlockBlob',
-          },
-        });
+        containerClient.uploadBlockBlob(targetFileName, content, content.length);
       }
     } finally {
       await rm(tempFolder, {