+++ workspaces/packages/bicep-assets/src/utils/manifest.ts
@@ -1,7 +1,10 @@
 import * as t from 'typanion';
 
 export const isManifest = t.isObject({
-  resourceProviderId: t.isString(),
+  resourceProviderId: t.isOptional(t.isString()),
+  storageAccountName: t.isOptional(t.isString()),
+  subscription: t.isOptional(t.isString()),
+  resourceGroup: t.isOptional(t.isString()),
   assets: t.isRecord(t.isString()),
 });